@@ -19,7 +19,7 @@ dependencies = [
     "rich>=13.7.0",
     "types-requests>=2.32.4.20250611",
 ]
-requires-python = ">=3.9"
+requires-python = ">=3.11"
 readme = "README.md"
 license = {text = "MIT"}
 
@@ -46,7 +46,7 @@ build-backend = "hatchling.build"
 packages = ["backend"]
 
 [tool.ruff]
-target-version = "py39"
+target-version = "py311"
 line-length = 88
 
 [tool.ruff.lint]
@@ -88,7 +88,7 @@ extend-exclude = '''
 '''
 
 [tool.mypy]
-python_version = "3.9"
+python_version = "3.11"
 check_untyped_defs = true
 disallow_any_generics = true
 disallow_incomplete_defs = true
